🔨 [TEA-60]: Refactored docker-compose and eslint
