--- .env.sample
@@ -1,5 +1,0 @@
-PUBLIC_URL=http://localhost:3000
-
-DB_USER=secret
-DB_PASSWORD=admin
-DB_NAME=communitea-db
