+++ docker-compose.yaml
@@ -30,8 +30,8 @@ services:
     image: node:lts-bullseye
     # Change the run dev command to run build for production
     command: bash -c "npm install && npm run dev"
-    working_dir: /app
+    working_dir: /frontend
     volumes:
-      - ./app:/app
+      - ./frontend:/frontend
     ports:
       - 3000:3000
