+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import LandingPage from './components/Landing Page/LandingPage';
+import LandingPage from './components/LandingPage/LandingPage';
 import CommuniTeaPage from './components/CommuniTEA/CommuniTeaPage';
 import EventDetailPage from './components/IndividualEvent/EventDeatailPage';
