+++ frontend/src/components/CommuniTEA/CommuniTeaPage.tsx
@@ -1,5 +1,5 @@
-import Footer from '../Landing Page/Footer';
-import NavBar from '../Landing Page/Navbar';
+import Footer from '../LandingPage/Footer';
+import NavBar from '../LandingPage/Navbar';
 import Filters from './Filters';
 import LocationSearch from './LocationSearch';
 import LocalBusinesses from './LocalBusinesses';
